add restrictions checks to participant interface dragging
