+++ components/participant_interface_main/setup.mjs
@@ -1,3 +1,4 @@
+import { getAnswerStr } from "./graph_handlers.mjs";
 import { handleInjectionMessage, setAnswer } from "./iframe_coms.mjs";
 
 // handle messages from the Qualtrics injection
@@ -5,6 +6,6 @@ window.addEventListener("message", handleInjectionMessage);
 
 // submit the answer on button press
 document.getElementById("submitButton").addEventListener("click", () => {
-  let answerStr = getAnswerStr(); // defined in `participant_interface.js`
+  let answerStr = getAnswerStr();
   setAnswer(answerStr);
 });
